generates random 8 digit number with a message
you will get unique messages based on what kind of number you get. it will depend on how many numbers in a row are matching at the end of the number. There are singles, doubles, triples, and on to octuples. Singles are the most common, and less common the farther on you go. Getting something like octuples is EXTREMELY rare and you must be proud if you get it.